// ui/src/api/generated/model/productCategoryListItemResponse.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Whirlpool Inspection Tool API
 * Whirlpool Inspection Tool API APIs developed by Scopt Analytics.
 * OpenAPI spec version: 1.0.3
 */

export interface ProductCategoryListItemResponse {
  id: number;
  uuid: string;
  name: string;
  category_type: string;
  sub_category_type: string;
  category_code: string;
  category_description: string;
  is_active: boolean;
  created_at: string;
  updated_at: string;
  products_count: number;
  inspections_count: number;
  inspection_inbound_under_review: number;
  inspection_outbound_under_review: number;
  inspection_inbound_approved: number;
  inspection_outbound_approved: number;
  inspection_inbound_rejected: number;
  inspection_outbound_rejected: number;
}